backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
beanie==1.26.0
motor==3.4.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
firebase-admin==6.5.0
qrcode==7.4.2
Pillow==10.4.0
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
# Pin bcrypt to a version compatible with passlib to avoid __about__ / password length issues
bcrypt==3.2.2
pymongo<4.9.0
slowapi==0.1.9
SQLAlchemy==2.0.36
boto3==1.42.4
python-socketio==5.11.0
APScheduler==3.10.4
httpx==0.27.2
